requirements.txt
python-dotenv==1.0.0
llama-index==0.10.1
openai==1.3.0
psutil==5.9.0
youtube_transcript_api==0.6.1
torch==2.1.0
numpy==1.24.3